Use f32 for typing_cps and require Silero VAD
Convert CODESCRIBE_TYPING_CPS to f32 across the codebase so fractional typing CPS values are supported (settings, loader parsing, and migrate). Add an end-to-end test to verify decimal persistence. Make the local Silero VAD required (non-optional) in ConversationEngine: fail-fast on init, call feed() directly for speech probability, and reset the VAD on conversation reset. Fix UI bootstrap to use owned Strings (avoid Box::leak) and pass string references to add_row. These changes ensure correct numeric handling for typing CPS and enforce the presence of the Silero VAD for conversation mode.[ok-commit]

+++ core/config/migrate.rs
@@ -134,7 +134,7 @@ pub fn migrate_if_needed() {
         settings.buffer_delay_ms = Some(n);
     }
     if let Ok(v) = std::env::var("CODESCRIBE_TYPING_CPS")
-        && let Ok(n) = v.parse::<u64>()
+        && let Ok(n) = v.parse::<f32>()
     {
         settings.typing_cps = Some(n);
     }
